Fixed issue !194.
Changed the order of if-statements to solve issue 194.

The problem was caused because the user provided an content encoding type.

Every time if an content encoding type was given then the create method of 'Subscription' returned a new subscription based on the 'authToken' authorization format.

The issue creator used the variant where you provide the array 'keys' for authorization.

This format could not be used if you provided an content encoding type.

This bug is now fixed by changing the if-statement order.